+++ tests/beacon/state_machines/test_proposer_signature_validation.py
@@ -4,7 +4,7 @@ from eth_utils import (
     ValidationError,
 )
 
-from eth._utils import bls
+from eth2._utils import bls
 
 from eth.beacon.constants import (
     GWEI_PER_ETH,
