+++ tests/beacon/test_aggregation.py
@@ -5,7 +5,7 @@ from hypothesis import (
     strategies as st,
 )
 
-from eth._utils import bls
+from eth2._utils import bls
 from eth._utils.bitfield import (
     get_empty_bitfield,
     has_voted,
